+++ driver/js/packages/hippy-vue/src/renderer/native/style/__tests__/index.test.js
@@ -1,3 +1,23 @@
+/*
+ * Tencent is pleased to support the open source community by making
+ * Hippy available.
+ *
+ * Copyright (C) 2017-2022 THL A29 Limited, a Tencent company.
+ * All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 import test, { before } from 'ava';
 import { fromAstNodes, SelectorsMap } from '../index';
 import ElementNode from '../../../element-node';
@@ -37,6 +57,7 @@ test('AppendSelector test', (t) => {
     selectors: [
       '#id',
       '*',
+      '#id[attr="test"]',
     ],
     declarations: [
       {
@@ -49,6 +70,11 @@ test('AppendSelector test', (t) => {
         property: 'UniversalSelector',
         value: 'UniversalSelector',
       },
+      {
+        type: 'declaration',
+        property: 'AttributeSelector',
+        value: 'AttributeSelector',
+      },
     ],
   }];
   const appendRules = fromAstNodes(APPEND_AST);
@@ -56,5 +82,5 @@ test('AppendSelector test', (t) => {
   const node = new ElementNode('div');
   node.setAttribute('id', 'id');
   const matchedCSS = cssMap.query(node);
-  t.is(matchedCSS.selectors.length, 9);
+  t.is(matchedCSS.selectors.length, 10);
 });
